pyproject.toml
[project]
name = "cq-20250513"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyascii==0.3.2",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "authlib==1.5.2",
    "bandit==1.8.3",
    "beautifulsoup4==4.13.4",
    "black==25.1.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "coverage==7.8.0",
    "crispy-tailwind==1.0.3",
    "cryptography==44.0.3",
    "defusedxml==0.7.1",
    "django==5.2.1",
    "django-appconf==1.1.0",
    "django-cities-light>=3.10.1",
    "django-components==0.139.1",
    "django-compressor==4.5.1",
    "django-countries>=7.6.1",
    "django-crispy-forms==2.4",
    "django-debug-toolbar==5.2.0",
    "django-environ==0.12.0",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-modelcluster==6.4",
    "django-permissionedforms==0.1",
    "django-pytest>=0.2.0",
    "django-stubs-ext==5.2.0",
    "django-stubs[compatible-mypy]==5.2.0",
    "django-taggit==6.1.0",
    "django-tasks==0.7.0",
    "django-threadlocals>=0.10",
    "django-treebeard==4.7.1",
    "djangorestframework==3.16.0",
    "djc-core-html-parser==1.0.2",
    "dparse==0.6.4",
    "draftjs-exporter==5.1.0",
    "et-xmlfile==2.0.0",
    "factory-boy==3.3.3",
    "faker==37.3.0",
    "filelock==3.16.1",
    "filetype==1.2.0",
    "flake8==7.2.0",
    "folium>=0.20.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "isort==6.0.1",
    "jinja2==3.1.6",
    "joblib==1.5.0",
    "laces==0.1.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==4.0.0",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "mypy==1.15.0",
    "mypy-extensions==1.1.0",
    "nltk==3.9.1",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pbr==6.1.1",
    "pillow==11.2.1",
    "pillow-heif==0.21.0",
    "platformdirs==4.3.8",
    "pluggy==1.5.0",
    "psutil==6.1.1",
    "py-cpuinfo==9.0.0",
    "pycodestyle==2.13.0",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pyflakes==3.3.2",
    "pygments==2.19.1",
    "pytest==8.3.5",
    "pytest-benchmark==5.1.0",
    "pytest-cov==6.1.1",
    "pytest-django==4.11.1",
    "pytest-playwright",
    "pyyaml==6.0.2",
    "rcssmin==1.1.2",
    "regex==2024.11.6",
    "reportlab>=4.4.1",
    "requests==2.32.3",
    "rich==14.0.0",
    "rjsmin==1.2.2",
    "ruamel-yaml==0.18.10",
    "ruamel-yaml-clib==0.2.12",
    "safety==3.5.0",
    "safety-schemas==0.0.14",
    "setuptools==80.4.0",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlparse==0.5.3",
    "stevedore==5.4.1",
    "tailwind==3.1.5b0",
    "telepath==0.3.1",
    "tenacity==9.1.2",
    "tomlkit==0.13.2",
    "tqdm==4.67.1",
    "typer==0.15.3",
    "types-pyyaml==6.0.12.20250516",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wagtail==7.0",
    "willow==1.10.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "black>=25.1.0",
    "django-debug-toolbar>=5.2.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-html>=4.1.1",
]

[tool.black]
line_length = 120
target_version = ['py38']

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.dev"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
testpaths = ["tests"]
addopts = "--reuse-db"
markers = "django_db: marks tests as requiring database access"
